Allow Draft Sessions to Survive Editing/Disable Drafts for Cloning
This basically discerns if an edit session can use Drafts (e.g Cloning can not) and we keep Draft sessions separated from normal advance sessions. Reason for that is we need to be able to Edit a Node without touching an ongoing/Draft session too.

We use the same widget ID as temporary private store storage but we add a -draft

Also:
- Marks Draft as Saved (message( with normal Operations (next/previous buttons) by setting  $form_state->set('draft_saved', TRUE); Those buttons still validate the form v/s the Top navigation one.
- Add the Step number in the message when reaching the last step and validation for previous ones fails

+++ src/EventSubscriber/WebformStrawberryfieldDeleteTmpStorage.php
@@ -111,8 +111,13 @@ class WebformStrawberryfieldDeleteTmpStorage implements EventSubscriberInterface
       foreach ($field->getValue() as $delta => $value) {
         $keyid = $this->getTempStoreKeyName($fieldname, $delta, '');
         $tempstore->delete($keyid);
-        // Delete also any cached errors
-        $tempstore->delete($keyid.'-errors');
+        // Delete also any cached errors and drafts if the autosave session
+        // Generated this ADO. If not (e.g Clone we will have marked that
+        // ADO's UUID as not autosave so we do not delete another ongoing session
+        if ($this->tempStoreFactory->get('archipel_autosave')->get($entity->uuid())) {
+          $tempstore->delete($keyid . '-errors');
+          $tempstore->delete($keyid . '-draft');
+        }
       }
     }
